backend/requirements.txt
# Build dependencies (must be first)
setuptools>=65.0.0
wheel>=0.38.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# HTTP Requests
requests>=2.28.0

# Video Processing
moviepy==1.0.3
opencv-python-headless>=4.8.0
scenedetect>=0.6.0
imageio>=2.25.0
imageio-ffmpeg>=0.4.8

# Audio/Video Analysis
numpy>=1.24.0

# Data Validation
pydantic>=2.0.0

# System utilities
python-multipart>=0.0.5

# agent
google-adk

# Google Cloud dependencies
google-cloud-storage>=2.10.0
google-auth>=2.20.0

# Environment management
python-dotenv>=1.0.0

# Database dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter
aiosqlite>=0.19.0       # SQLite adapter for development

# Authentication dependencies
PyJWT>=2.8.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
passlib[bcrypt]>=1.7.4
google-generativeai
pydub 
google-cloud-speech 
google-cloud-translate 
google-cloud-texttospeech